@@ -15,21 +15,24 @@ import android.view.View;
 import android.widget.FrameLayout;
 import android.widget.ImageView;
 
-import com.bumptech.glide.DrawableRequestBuilder;
-import com.bumptech.glide.GenericRequestBuilder;
 import com.bumptech.glide.Glide;
+import com.bumptech.glide.RequestBuilder;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners;
+import com.bumptech.glide.request.RequestOptions;
 
 import org.thoughtcrime.securesms.R;
 import org.thoughtcrime.securesms.crypto.MasterSecret;
 import org.thoughtcrime.securesms.database.AttachmentDatabase;
 import org.thoughtcrime.securesms.mms.DecryptableStreamUriLoader.DecryptableUri;
-import org.thoughtcrime.securesms.mms.RoundedCorners;
+import org.thoughtcrime.securesms.mms.GlideApp;
 import org.thoughtcrime.securesms.mms.Slide;
 import org.thoughtcrime.securesms.mms.SlideClickListener;
 import org.thoughtcrime.securesms.util.Util;
 import org.thoughtcrime.securesms.util.ViewUtil;
 import org.whispersystems.libsignal.util.guava.Optional;
+
+import static com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions.withCrossFade;
 
 public class ThumbnailView extends FrameLayout {
 
@@ -141,18 +144,18 @@ public class ThumbnailView extends FrameLayout {
 
     if      (slide.getThumbnailUri() != null) buildThumbnailGlideRequest(slide, masterSecret).into(image);
     else if (slide.hasPlaceholder())          buildPlaceholderGlideRequest(slide).into(image);
-    else                                      Glide.clear(image);
+    else                                      Glide.with(getContext()).clear(image);
   }
 
   public void setImageResource(@NonNull MasterSecret masterSecret, @NonNull Uri uri) {
     if (transferControls.isPresent()) getTransferControls().setVisibility(View.GONE);
 
-    Glide.with(getContext())
-         .load(new DecryptableUri(masterSecret, uri))
-         .diskCacheStrategy(DiskCacheStrategy.NONE)
-         .crossFade()
-         .transform(new RoundedCorners(getContext(), true, radius, backgroundColorHint))
-         .into(image);
+    GlideApp.with(getContext())
+            .load(new DecryptableUri(masterSecret, uri))
+            .diskCacheStrategy(DiskCacheStrategy.NONE)
+            .transform(new RoundedCorners(radius))
+            .transition(withCrossFade())
+            .into(image);
   }
 
   public void setThumbnailClickListener(SlideClickListener listener) {
@@ -164,7 +167,7 @@ public class ThumbnailView extends FrameLayout {
   }
 
   public void clear() {
-    if (isContextValid())             Glide.clear(image);
+    if (isContextValid())             Glide.with(getContext()).clear(image);
     if (transferControls.isPresent()) getTransferControls().clear();
 
     slide = null;
@@ -181,24 +184,23 @@ public class ThumbnailView extends FrameLayout {
            !((Activity)getContext()).isDestroyed();
   }
 
-  private GenericRequestBuilder buildThumbnailGlideRequest(@NonNull Slide slide, @NonNull MasterSecret masterSecret) {
-    @SuppressWarnings("ConstantConditions")
-    DrawableRequestBuilder<DecryptableUri> builder = Glide.with(getContext())
-                                                          .load(new DecryptableUri(masterSecret, slide.getThumbnailUri()))
-                                                          .diskCacheStrategy(DiskCacheStrategy.NONE)
-                                                          .crossFade()
-                                                          .transform(new RoundedCorners(getContext(), true, radius, backgroundColorHint));
+  private RequestBuilder buildThumbnailGlideRequest(@NonNull Slide slide, @NonNull MasterSecret masterSecret) {
+    RequestBuilder builder = GlideApp.with(getContext())
+                                     .load(new DecryptableUri(masterSecret, slide.getThumbnailUri()))
+                                     .diskCacheStrategy(DiskCacheStrategy.NONE)
+                                     .transform(new RoundedCorners(radius))
+                                     .transition(withCrossFade());
 
     if (slide.isInProgress()) return builder;
-    else                      return builder.error(R.drawable.ic_missing_thumbnail_picture);
+    else                      return builder.apply(RequestOptions.errorOf(R.drawable.ic_missing_thumbnail_picture));
   }
 
-  private GenericRequestBuilder buildPlaceholderGlideRequest(Slide slide) {
-    return Glide.with(getContext())
-                .load(slide.getPlaceholderRes(getContext().getTheme()))
-                .asBitmap()
-                .diskCacheStrategy(DiskCacheStrategy.NONE)
-                .fitCenter();
+  private RequestBuilder buildPlaceholderGlideRequest(Slide slide) {
+    return GlideApp.with(getContext())
+                   .asBitmap()
+                   .load(slide.getPlaceholderRes(getContext().getTheme()))
+                   .diskCacheStrategy(DiskCacheStrategy.NONE)
+                   .fitCenter();
   }
 
   private class ThumbnailClickDispatcher implements View.OnClickListener {
