@@ -1,7 +1,6 @@
 package org.thoughtcrime.securesms.components;
 
 import android.content.Context;
-import android.graphics.Canvas;
 import android.net.Uri;
 import android.os.AsyncTask;
 import android.support.annotation.Nullable;
@@ -10,37 +9,33 @@ import android.util.Log;
 import android.util.Pair;
 import android.view.View;
 import android.widget.FrameLayout;
-import android.widget.ImageView;
 
-import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
-import com.bumptech.glide.load.resource.drawable.GlideDrawable;
-import com.bumptech.glide.request.RequestListener;
-import com.bumptech.glide.request.target.GlideDrawableImageViewTarget;
 import com.bumptech.glide.request.target.Target;
 import com.davemorrissey.labs.subscaleview.ImageSource;
 import com.davemorrissey.labs.subscaleview.SubsamplingScaleImageView;
+import com.github.chrisbanes.photoview.PhotoView;
 
 import org.thoughtcrime.securesms.R;
 import org.thoughtcrime.securesms.components.subsampling.AttachmentBitmapDecoder;
 import org.thoughtcrime.securesms.components.subsampling.AttachmentRegionDecoder;
 import org.thoughtcrime.securesms.crypto.MasterSecret;
 import org.thoughtcrime.securesms.mms.DecryptableStreamUriLoader.DecryptableUri;
+import org.thoughtcrime.securesms.mms.GlideApp;
 import org.thoughtcrime.securesms.mms.PartAuthority;
 import org.thoughtcrime.securesms.util.BitmapDecodingException;
 import org.thoughtcrime.securesms.util.BitmapUtil;
+import org.thoughtcrime.securesms.util.MediaUtil;
 
 import java.io.IOException;
 import java.io.InputStream;
 
-import uk.co.senab.photoview.PhotoViewAttacher;
 
 public class ZoomingImageView extends FrameLayout {
 
   private static final String TAG = ZoomingImageView.class.getName();
 
-  private final ImageView                 imageView;
-  private final PhotoViewAttacher         imageViewAttacher;
+  private final PhotoView                 photoView;
   private final SubsamplingScaleImageView subsamplingImageView;
 
   public ZoomingImageView(Context context) {
@@ -56,9 +51,8 @@ public class ZoomingImageView extends FrameLayout {
 
     inflate(context, R.layout.zooming_image_view, this);
 
-    this.imageView            = (ImageView) findViewById(R.id.image_view);
-    this.subsamplingImageView = (SubsamplingScaleImageView) findViewById(R.id.subsampling_image_view);
-    this.imageViewAttacher     = new PhotoViewAttacher(imageView);
+    this.photoView            = findViewById(R.id.image_view);
+    this.subsamplingImageView = findViewById(R.id.subsampling_image_view);
 
     this.subsamplingImageView.setBitmapDecoderClass(AttachmentBitmapDecoder.class);
     this.subsamplingImageView.setRegionDecoderClass(AttachmentRegionDecoder.class);
@@ -74,7 +68,7 @@ public class ZoomingImageView extends FrameLayout {
     new AsyncTask<Void, Void, Pair<Integer, Integer>>() {
       @Override
       protected @Nullable Pair<Integer, Integer> doInBackground(Void... params) {
-        if (contentType.equals("image/gif")) return null;
+        if (MediaUtil.isGif(contentType)) return null;
 
         try {
           InputStream inputStream = PartAuthority.getAttachmentStream(context, masterSecret, uri);
@@ -100,32 +94,26 @@ public class ZoomingImageView extends FrameLayout {
   }
 
   private void setImageViewUri(MasterSecret masterSecret, Uri uri) {
+    photoView.setVisibility(View.VISIBLE);
     subsamplingImageView.setVisibility(View.GONE);
-    imageView.setVisibility(View.VISIBLE);
 
-    Glide.with(getContext())
-         .load(new DecryptableUri(masterSecret, uri))
-         .diskCacheStrategy(DiskCacheStrategy.NONE)
-         .dontTransform()
-         .dontAnimate()
-         .into(new GlideDrawableImageViewTarget(imageView) {
-           @Override protected void setResource(GlideDrawable resource) {
-             super.setResource(resource);
-             imageViewAttacher.update();
-           }
-         });
+    GlideApp.with(getContext())
+            .load(new DecryptableUri(masterSecret, uri))
+            .diskCacheStrategy(DiskCacheStrategy.NONE)
+            .dontTransform()
+            .override(Target.SIZE_ORIGINAL, Target.SIZE_ORIGINAL)
+            .into(photoView);
   }
 
   private void setSubsamplingImageViewUri(Uri uri) {
     subsamplingImageView.setVisibility(View.VISIBLE);
-    imageView.setVisibility(View.GONE);
+    photoView.setVisibility(View.GONE);
 
     subsamplingImageView.setImage(ImageSource.uri(uri));
   }
 
-
   public void cleanup() {
-    imageView.setImageDrawable(null);
+    photoView.setImageDrawable(null);
     subsamplingImageView.recycle();
   }
 }
