@@ -3,7 +3,9 @@ package org.thoughtcrime.securesms.giph.ui;
 
 import android.content.Context;
 import android.graphics.drawable.ColorDrawable;
+import android.graphics.drawable.Drawable;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -11,16 +13,18 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ProgressBar;
 
-import com.bumptech.glide.DrawableRequestBuilder;
 import com.bumptech.glide.Glide;
+import com.bumptech.glide.RequestBuilder;
+import com.bumptech.glide.load.DataSource;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
-import com.bumptech.glide.load.resource.drawable.GlideDrawable;
+import com.bumptech.glide.load.engine.GlideException;
 import com.bumptech.glide.request.RequestListener;
 import com.bumptech.glide.request.target.Target;
 
 import org.thoughtcrime.securesms.R;
 import org.thoughtcrime.securesms.color.MaterialColor;
 import org.thoughtcrime.securesms.giph.model.GiphyImage;
+import org.thoughtcrime.securesms.mms.GlideApp;
 import org.thoughtcrime.securesms.util.Util;
 import org.thoughtcrime.securesms.util.ViewUtil;
 
@@ -37,7 +41,7 @@ public class GiphyAdapter extends RecyclerView.Adapter<GiphyAdapter.GiphyViewHol
   private Context              context;
   private OnItemClickListener  listener;
 
-  class GiphyViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener, RequestListener<String, GlideDrawable> {
+  class GiphyViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener, RequestListener<Drawable> {
 
     public AspectRatioImageView thumbnail;
     public GiphyImage           image;
@@ -58,7 +62,7 @@ public class GiphyAdapter extends RecyclerView.Adapter<GiphyAdapter.GiphyViewHol
     }
 
     @Override
-    public boolean onException(Exception e, String model, Target<GlideDrawable> target, boolean isFirstResource) {
+    public boolean onLoadFailed(@Nullable GlideException e, Object model, Target<Drawable> target, boolean isFirstResource) {
       Log.w(TAG, e);
 
       synchronized (this) {
@@ -69,10 +73,11 @@ public class GiphyAdapter extends RecyclerView.Adapter<GiphyAdapter.GiphyViewHol
       }
 
       return false;
+
     }
 
     @Override
-    public boolean onResourceReady(GlideDrawable resource, String model, Target<GlideDrawable> target, boolean isFromMemoryCache, boolean isFirstResource) {
+    public boolean onResourceReady(Drawable resource, Object model, Target<Drawable> target, DataSource dataSource, boolean isFirstResource) {
       synchronized (this) {
         if (image.getGifUrl().equals(model)) {
           this.modelReady = true;
@@ -82,6 +87,7 @@ public class GiphyAdapter extends RecyclerView.Adapter<GiphyAdapter.GiphyViewHol
 
       return false;
     }
+
 
     public File getFile(boolean forMms) throws ExecutionException, InterruptedException {
       synchronized (this) {
@@ -103,7 +109,7 @@ public class GiphyAdapter extends RecyclerView.Adapter<GiphyAdapter.GiphyViewHol
   }
 
   GiphyAdapter(Context context, List<GiphyImage> images) {
-    this.context = context;
+    this.context = context.getApplicationContext();
     this.images  = images;
   }
 
@@ -134,32 +140,33 @@ public class GiphyAdapter extends RecyclerView.Adapter<GiphyAdapter.GiphyViewHol
     holder.thumbnail.setAspectRatio(image.getGifAspectRatio());
     holder.gifProgress.setVisibility(View.GONE);
 
-    DrawableRequestBuilder<String> thumbnailRequest = Glide.with(context)
-                                                           .load(image.getStillUrl());
+    RequestBuilder<Drawable> thumbnailRequest = GlideApp.with(context)
+                                                        .load(image.getStillUrl())
+                                                        .diskCacheStrategy(DiskCacheStrategy.ALL);
 
     if (Util.isLowMemory(context)) {
-      Glide.with(context)
-           .load(image.getStillUrl())
-           .placeholder(new ColorDrawable(Util.getRandomElement(MaterialColor.values()).toConversationColor(context)))
-           .diskCacheStrategy(DiskCacheStrategy.ALL)
-           .into(holder.thumbnail);
+      GlideApp.with(context)
+              .load(image.getStillUrl())
+              .placeholder(new ColorDrawable(Util.getRandomElement(MaterialColor.values()).toConversationColor(context)))
+              .diskCacheStrategy(DiskCacheStrategy.ALL)
+              .into(holder.thumbnail);
 
       holder.setModelReady();
     } else {
-      Glide.with(context)
-           .load(image.getGifUrl())
-           .thumbnail(thumbnailRequest)
-           .placeholder(new ColorDrawable(Util.getRandomElement(MaterialColor.values()).toConversationColor(context)))
-           .diskCacheStrategy(DiskCacheStrategy.ALL)
-           .listener(holder)
-           .into(holder.thumbnail);
+      GlideApp.with(context)
+              .load(image.getGifUrl())
+              .thumbnail(thumbnailRequest)
+              .placeholder(new ColorDrawable(Util.getRandomElement(MaterialColor.values()).toConversationColor(context)))
+              .diskCacheStrategy(DiskCacheStrategy.ALL)
+              .listener(holder)
+              .into(holder.thumbnail);
     }
   }
 
   @Override
   public void onViewRecycled(GiphyViewHolder holder) {
     super.onViewRecycled(holder);
-    Glide.clear(holder.thumbnail);
+    Glide.with(context).clear(holder.thumbnail);
   }
 
   @Override
