@@ -1,6 +1,9 @@
 package org.thoughtcrime.securesms.glide;
 
+import android.support.annotation.NonNull;
+
 import com.bumptech.glide.Priority;
+import com.bumptech.glide.load.DataSource;
 import com.bumptech.glide.load.data.DataFetcher;
 import com.bumptech.glide.load.model.GlideUrl;
 import com.bumptech.glide.util.ContentLengthInputStream;
@@ -17,7 +20,7 @@ import okhttp3.ResponseBody;
 /**
  * Fetches an {@link InputStream} using the okhttp library.
  */
-public class OkHttpStreamFetcher implements DataFetcher<InputStream> {
+class OkHttpStreamFetcher implements DataFetcher<InputStream> {
 
   private static final String TAG = OkHttpStreamFetcher.class.getName();
 
@@ -26,33 +29,38 @@ public class OkHttpStreamFetcher implements DataFetcher<InputStream> {
   private       InputStream  stream;
   private       ResponseBody responseBody;
 
-  public OkHttpStreamFetcher(OkHttpClient client, GlideUrl url) {
+  OkHttpStreamFetcher(OkHttpClient client, GlideUrl url) {
     this.client = client;
     this.url = url;
   }
 
   @Override
-  public InputStream loadData(Priority priority) throws Exception {
-    Request.Builder requestBuilder = new Request.Builder()
-        .url(url.toStringUrl());
+  public void loadData(Priority priority, DataCallback<? super InputStream> callback) {
+    try {
+      Request.Builder requestBuilder = new Request.Builder()
+          .url(url.toStringUrl());
 
-    for (Map.Entry<String, String> headerEntry : url.getHeaders().entrySet()) {
-      String key = headerEntry.getKey();
-      requestBuilder.addHeader(key, headerEntry.getValue());
+      for (Map.Entry<String, String> headerEntry : url.getHeaders().entrySet()) {
+        String key = headerEntry.getKey();
+        requestBuilder.addHeader(key, headerEntry.getValue());
+      }
+
+      Request  request  = requestBuilder.build();
+      Response response = client.newCall(request).execute();
+
+      responseBody = response.body();
+
+      if (!response.isSuccessful()) {
+        throw new IOException("Request failed with code: " + response.code());
+      }
+
+      long contentLength = responseBody.contentLength();
+      stream = ContentLengthInputStream.obtain(responseBody.byteStream(), contentLength);
+
+      callback.onDataReady(stream);
+    } catch (IOException e) {
+      callback.onLoadFailed(e);
     }
-
-    Request  request  = requestBuilder.build();
-    Response response = client.newCall(request).execute();
-
-    responseBody = response.body();
-
-    if (!response.isSuccessful()) {
-      throw new IOException("Request failed with code: " + response.code());
-    }
-
-    long contentLength = responseBody.contentLength();
-    stream = ContentLengthInputStream.obtain(responseBody.byteStream(), contentLength);
-    return stream;
   }
 
   @Override
@@ -70,12 +78,19 @@ public class OkHttpStreamFetcher implements DataFetcher<InputStream> {
   }
 
   @Override
-  public String getId() {
-    return url.getCacheKey();
-  }
-
-  @Override
   public void cancel() {
     // TODO: call cancel on the client when this method is called on a background thread. See #257
+  }
+
+  @NonNull
+  @Override
+  public Class<InputStream> getDataClass() {
+    return InputStream.class;
+  }
+
+  @NonNull
+  @Override
+  public DataSource getDataSource() {
+    return DataSource.REMOTE;
   }
 }
