@@ -1,12 +1,12 @@
 package org.thoughtcrime.securesms.glide;
 
-import android.content.Context;
+import android.support.annotation.Nullable;
 
-import com.bumptech.glide.load.data.DataFetcher;
-import com.bumptech.glide.load.model.GenericLoaderFactory;
+import com.bumptech.glide.load.Options;
 import com.bumptech.glide.load.model.GlideUrl;
 import com.bumptech.glide.load.model.ModelLoader;
 import com.bumptech.glide.load.model.ModelLoaderFactory;
+import com.bumptech.glide.load.model.MultiModelLoaderFactory;
 
 import org.thoughtcrime.securesms.giph.net.GiphyProxySelector;
 
@@ -19,9 +19,23 @@ import okhttp3.OkHttpClient;
  */
 public class OkHttpUrlLoader implements ModelLoader<GlideUrl, InputStream> {
 
-  /**
-   * The default factory for {@link OkHttpUrlLoader}s.
-   */
+  private final OkHttpClient client;
+
+  private OkHttpUrlLoader(OkHttpClient client) {
+    this.client = client;
+  }
+
+  @Nullable
+  @Override
+  public LoadData<InputStream> buildLoadData(GlideUrl glideUrl, int width, int height, Options options) {
+    return new LoadData<>(glideUrl, new OkHttpStreamFetcher(client, glideUrl));
+  }
+
+  @Override
+  public boolean handles(GlideUrl glideUrl) {
+    return true;
+  }
+
   public static class Factory implements ModelLoaderFactory<GlideUrl, InputStream> {
     private static volatile OkHttpClient internalClient;
     private OkHttpClient client;
@@ -39,22 +53,16 @@ public class OkHttpUrlLoader implements ModelLoader<GlideUrl, InputStream> {
       return internalClient;
     }
 
-    /**
-     * Constructor for a new Factory that runs requests using a static singleton client.
-     */
     public Factory() {
       this(getInternalClient());
     }
 
-    /**
-     * Constructor for a new Factory that runs requests using given client.
-     */
     private Factory(OkHttpClient client) {
       this.client = client;
     }
 
     @Override
-    public ModelLoader<GlideUrl, InputStream> build(Context context, GenericLoaderFactory factories) {
+    public ModelLoader<GlideUrl, InputStream> build(MultiModelLoaderFactory multiFactory) {
       return new OkHttpUrlLoader(client);
     }
 
@@ -62,16 +70,5 @@ public class OkHttpUrlLoader implements ModelLoader<GlideUrl, InputStream> {
     public void teardown() {
       // Do nothing, this instance doesn't own the client.
     }
-  }
-
-  private final OkHttpClient client;
-
-  private OkHttpUrlLoader(OkHttpClient client) {
-    this.client = client;
-  }
-
-  @Override
-  public DataFetcher<InputStream> getResourceFetcher(GlideUrl model, int width, int height) {
-    return new OkHttpStreamFetcher(client, model);
   }
 }
