@@ -17,15 +17,15 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 
-public class DecryptableStreamLocalUriFetcher extends StreamLocalUriFetcher {
+class DecryptableStreamLocalUriFetcher extends StreamLocalUriFetcher {
 
   private static final String TAG = DecryptableStreamLocalUriFetcher.class.getSimpleName();
 
   private Context      context;
   private MasterSecret masterSecret;
 
-  public DecryptableStreamLocalUriFetcher(Context context, MasterSecret masterSecret, Uri uri) {
-    super(context, uri);
+  DecryptableStreamLocalUriFetcher(Context context, MasterSecret masterSecret, Uri uri) {
+    super(context.getContentResolver(), uri);
     this.context      = context;
     this.masterSecret = masterSecret;
   }
