@@ -3,33 +3,49 @@ package org.thoughtcrime.securesms.mms;
 import android.content.Context;
 import android.net.Uri;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 
-import com.bumptech.glide.load.data.DataFetcher;
-import com.bumptech.glide.load.model.GenericLoaderFactory;
+import com.bumptech.glide.load.Key;
+import com.bumptech.glide.load.Options;
 import com.bumptech.glide.load.model.ModelLoader;
 import com.bumptech.glide.load.model.ModelLoaderFactory;
-import com.bumptech.glide.load.model.stream.StreamModelLoader;
+import com.bumptech.glide.load.model.MultiModelLoaderFactory;
 
 import org.thoughtcrime.securesms.crypto.MasterSecret;
 import org.thoughtcrime.securesms.mms.DecryptableStreamUriLoader.DecryptableUri;
 
 import java.io.InputStream;
+import java.security.MessageDigest;
 
-/**
- * A {@link ModelLoader} for translating uri models into {@link InputStream} data. Capable of handling 'http',
- * 'https', 'android.resource', 'content', and 'file' schemes. Unsupported schemes will throw an exception in
- * {@link #getResourceFetcher(Uri, int, int)}.
- */
-public class DecryptableStreamUriLoader implements StreamModelLoader<DecryptableUri> {
+public class DecryptableStreamUriLoader implements ModelLoader<DecryptableUri, InputStream> {
+
   private final Context context;
 
-  /**
-   * THe default factory for {@link com.bumptech.glide.load.model.stream.StreamUriLoader}s.
-   */
-  public static class Factory implements ModelLoaderFactory<DecryptableUri, InputStream> {
+  private DecryptableStreamUriLoader(Context context) {
+    this.context = context;
+  }
+
+  @Nullable
+  @Override
+  public LoadData<InputStream> buildLoadData(DecryptableUri decryptableUri, int width, int height, Options options) {
+    return new LoadData<>(decryptableUri, new DecryptableStreamLocalUriFetcher(context, decryptableUri.masterSecret, decryptableUri.uri));
+  }
+
+  @Override
+  public boolean handles(DecryptableUri decryptableUri) {
+    return true;
+  }
+
+  static class Factory implements ModelLoaderFactory<DecryptableUri, InputStream> {
+
+    private final Context context;
+
+    Factory(Context context) {
+      this.context = context.getApplicationContext();
+    }
 
     @Override
-    public StreamModelLoader<DecryptableUri> build(Context context, GenericLoaderFactory factories) {
+    public ModelLoader<DecryptableUri, InputStream> build(MultiModelLoaderFactory multiFactory) {
       return new DecryptableStreamUriLoader(context);
     }
 
@@ -39,22 +55,18 @@ public class DecryptableStreamUriLoader implements StreamModelLoader<Decryptable
     }
   }
 
-  public DecryptableStreamUriLoader(Context context) {
-    this.context = context;
-  }
-
-  @Override
-  public DataFetcher<InputStream> getResourceFetcher(DecryptableUri model, int width, int height) {
-    return new DecryptableStreamLocalUriFetcher(context, model.masterSecret, model.uri);
-  }
-
-  public static class DecryptableUri {
+  public static class DecryptableUri implements Key {
     public @NonNull MasterSecret masterSecret;
     public @NonNull Uri          uri;
 
     public DecryptableUri(@NonNull MasterSecret masterSecret, @NonNull Uri uri) {
       this.masterSecret = masterSecret;
       this.uri          = uri;
+    }
+
+    @Override
+    public void updateDiskCacheKey(MessageDigest messageDigest) {
+      messageDigest.update(uri.toString().getBytes());
     }
 
     @Override
